+++ Gear/Propeller/MemoryManager.cs
@@ -24,115 +24,175 @@ using Gear.EmulationCore;
 
 namespace Gear.Propeller
 {
+    /// <summary></summary>
     public class DirectMemory
     {
+        /// <summary></summary>
         protected byte[] Memory;
 
-        public byte this[int offset]
-        {
-            get
-            {
-                if (offset >= Memory.Length)
-                    return 0x55;
-                return Memory[offset];
-            }
-        }
+        /// <summary></summary>
+        public byte this[int offset] =>
+            offset >= Memory.Length ? (byte)0x55 : Memory[offset];
 
-        public byte DirectReadByte(uint address)
-        {
-            return Memory[address & PropellerCPU.MAX_RAM_ADDR];
-        }
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <returns></returns>
+        public byte DirectReadByte(uint address) =>
+            Memory[address & PropellerCPU.MAX_RAM_ADDR];
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <returns></returns>
+        /// @version v22.04.02 - Removed redundant increment.
         public ushort DirectReadWord(uint address)
         {
             address &= 0xFFFFFFFE;
-            return (ushort)(Memory[(address++) & PropellerCPU.MAX_RAM_ADDR] |
-                (Memory[(address++) & PropellerCPU.MAX_RAM_ADDR] << 8));
+            return (ushort)(Memory[address++ & PropellerCPU.MAX_RAM_ADDR] |
+                (Memory[address & PropellerCPU.MAX_RAM_ADDR] << 8));
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <returns></returns>
+        /// @version v22.04.02 - Removed redundant increment.
         public uint DirectReadLong(uint address)
         {
             address &= 0xFFFFFFFC;
-            return (uint)Memory[(address++) & PropellerCPU.MAX_RAM_ADDR]
-                | (uint)(Memory[(address++) & PropellerCPU.MAX_RAM_ADDR] << 8)
-                | (uint)(Memory[(address++) & PropellerCPU.MAX_RAM_ADDR] << 16)
-                | (uint)(Memory[(address++) & PropellerCPU.MAX_RAM_ADDR] << 24);
+            return (uint)Memory[address++ & PropellerCPU.MAX_RAM_ADDR]
+                | (uint)(Memory[address++ & PropellerCPU.MAX_RAM_ADDR] << 8)
+                | (uint)(Memory[address++ & PropellerCPU.MAX_RAM_ADDR] << 16)
+                | (uint)(Memory[address & PropellerCPU.MAX_RAM_ADDR] << 24);
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <param name="value"></param>
+        /// @version v22.04.02 - Removed redundant increment.
         public void DirectWriteByte(uint address, byte value)
         {
             if ((address & 0x8000) != 0)
                 return;
-            Memory[(address++) & 0x7FFF] = value;
+            Memory[address & 0x7FFF] = value;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <param name="value"></param>
+        /// @version v22.04.02 - Removed redundant increment.
         public void DirectWriteWord(uint address, ushort value)
         {
             address &= 0xFFFFFFFE;
             DirectWriteByte(address++, (byte) value);
-            DirectWriteByte(address++, (byte)(value >> 8));
+            DirectWriteByte(address, (byte)(value >> 8));
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="address"></param>
+        /// <param name="value"></param>
+        /// @version v22.04.02 - Removed redundant increment.
         public void DirectWriteLong(uint address, uint value)
         {
             address &= 0xFFFFFFFC;
             DirectWriteByte(address++, (byte)value);
             DirectWriteByte(address++, (byte)(value >> 8));
             DirectWriteByte(address++, (byte)(value >> 16));
-            DirectWriteByte(address++, (byte)(value >> 24));
+            DirectWriteByte(address, (byte)(value >> 24));
         }
     }
 
+    /// <summary>
+    ///
+    /// </summary>
     public class MemoryManager
     {
-        private readonly DirectMemory Memory;
+        /// <summary></summary>
+        private readonly DirectMemory _memory;
+        /// <summary></summary>
         public uint Address { get; protected set; }
 
-        public MemoryManager(DirectMemory Memory, uint Address = 0)
+        /// <summary>Default constructor.</summary>
+        /// <param name="memory"></param>
+        /// <param name="address"></param>
+        public MemoryManager(DirectMemory memory, uint address)
         {
-            this.Memory = Memory;
-            this.Address = Address;
-        }
-        public byte ReadByte()
-        {
-            return this.Memory.DirectReadByte(this.Address++);
+            _memory = memory;
+            Address = address;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <returns></returns>
+        public byte ReadByte() =>
+            _memory.DirectReadByte(Address++);
+
+        /// <summary>
+        ///
+        /// </summary>
+        /// <returns></returns>
         public ushort ReadWord()
         {
-            this.Address &= 0xFFFFFFFE;
-            ushort read = this.Memory.DirectReadWord(this.Address);
-            this.Address += 2;
+            Address &= 0xFFFFFFFE;
+            ushort read = _memory.DirectReadWord(Address);
+            Address += 2;
             return read;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <returns></returns>
         public uint ReadLong()
         {
-            this.Address &= 0xFFFFFFFC;
-            uint read = this.Memory.DirectReadLong(this.Address);
-            this.Address += 4;
+            Address &= 0xFFFFFFFC;
+            uint read = _memory.DirectReadLong(Address);
+            Address += 4;
             return read;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="value"></param>
         public void WriteByte(byte value)
         {
-            if ((this.Address & 0x8000) != 0)
+            if ((Address & 0x8000) != 0)
                 return;
-            this.Memory.DirectWriteByte(this.Address++, value);
+            _memory.DirectWriteByte(Address++, value);
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="value"></param>
         public void WriteWord(ushort value)
         {
-            this.Address &= 0xFFFFFFFE;
-            this.Memory.DirectWriteWord(this.Address, value);
-            this.Address += 2;
+            Address &= 0xFFFFFFFE;
+            _memory.DirectWriteWord(Address, value);
+            Address += 2;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="value"></param>
         public void WriteLong(uint value)
         {
-            this.Address &= 0xFFFFFFFC;
-            this.Memory.DirectWriteLong(this.Address, value);
-            this.Address += 4;
+            Address &= 0xFFFFFFFC;
+            _memory.DirectWriteLong(Address, value);
+            Address += 4;
         }
     }
 }